+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+## Unreleased
+
+- Add `get_object`, `require_object`, `get_secret_object` and `require_secret_object` APIs to Python
+  `config` module [#2959](https://github.com/pulumi/pulumi/pull/2959)
+
 ## 0.17.25 (2019-07-19)
 
 - Support for Dynamic Providers in Python [#2900](https://github.com/pulumi/pulumi/pull/2900)